Add JavaScribt functionality for basic pano viewing
This commit contains the actual logic, i.e. the correct calls to Three.js and a panning-/zooming module called OrbitControls.js

The plan is to take this as a starting point, refactor and refine all the code until next week, so ideally we have a codebase with the same functionality, but much better expandability in the future.