Include jQuery
Adding the current version of the jQuery library. It adds a very easy functionality to read in JSON files from a directory. This would otherwise be handled quite ugly so I think it makes sense to include the lib at this point already.
Oh and btw it is published under the MIT license as well of course.